requirements.txt
minemeld-core>=0.9.31b2
requests